Improve latency metrics

+++ backend/websockets/src/actors/mod.rs
@@ -7,7 +7,7 @@ pub mod messages;
 pub mod race;
 pub mod ws;
 
-pub fn measure_latency(msg: &LocationUpdateMessage) -> i64 {
+pub fn measure_latency(msg: &LocationUpdateMessage) -> f64 {
     let now = Utc::now().timestamp_millis();
-    now - (msg.timestamp as i64 * 1000)
+    (now - (msg.timestamp as i64 * 1000)) as f64
 }
